refactored screening page to fix the code climate issue
