+++ app/javascript/screenings/ScreeningPage.jsx
@@ -1,16 +1,9 @@
-import * as screeningActions from 'actions/screeningActions'
-import * as personCardActions from 'actions/personCardActions'
-import {setPageMode} from 'actions/screeningPageActions'
-import {fetchHistoryOfInvolvements, clearHistoryOfInvolvement} from 'actions/historyOfInvolvementActions'
-import {clearRelationships} from 'actions/relationshipsActions'
 import PersonCardView from 'screenings/PersonCardView'
 import PropTypes from 'prop-types'
 import React from 'react'
 import RelationshipsCardContainer from 'screenings/RelationshipsCardContainer'
 import CardContainer from 'containers/screenings/CardContainer'
 import {Link} from 'react-router'
-import {bindActionCreators} from 'redux'
-import {connect} from 'react-redux'
 import HistoryOfInvolvementContainer from 'containers/screenings/HistoryOfInvolvementContainer'
 import HistoryTableContainer from 'containers/screenings/HistoryTableContainer'
 import EmptyHistory from 'views/history/EmptyHistory'
@@ -32,15 +25,6 @@ import CrossReportShowContainer from 'containers/screenings/CrossReportShowConta
 import DecisionFormContainer from 'containers/screenings/DecisionFormContainer'
 import DecisionShowContainer from 'containers/screenings/DecisionShowContainer'
 import PageHeader from 'common/PageHeader'
-import {getScreeningSubmissionErrorsSelector, getApiValidationErrorsSelector} from 'selectors/errorsSelectors'
-import {selectParticipants} from 'selectors/participantSelectors'
-import {getScreeningTitleSelector, getScreeningIsReadOnlySelector} from 'selectors/screeningSelectors'
-import {
-  getAllCardsAreSavedValueSelector,
-  getScreeningHasErrorsSelector,
-  getPeopleHaveErrorsSelector,
-} from 'selectors/screening/screeningPageSelectors'
-import {selectStaffId} from 'selectors/userInfoSelectors'
 import {logEvent} from 'utils/analytics'
 
 const isDuplicatePerson = (participants, personOnScreening) => (
@@ -259,36 +243,3 @@ ScreeningPage.defaultProps = {
   mode: 'show',
 }
 
-export function mapStateToProps(state, _ownProps) {
-  return {
-    disableSubmitButton: !getAllCardsAreSavedValueSelector(state) ||
-      getScreeningHasErrorsSelector(state) ||
-      getPeopleHaveErrorsSelector(state),
-    editable: !getScreeningIsReadOnlySelector(state),
-    loaded: state.getIn(['screening', 'fetch_status']) === 'FETCHED',
-    mode: state.getIn(['screeningPage', 'mode']),
-    participants: selectParticipants(state).toJS(),
-    reference: state.getIn(['screening', 'reference']),
-    referralId: state.getIn(['screening', 'referral_id']),
-    hasApiValidationErrors: Boolean(getApiValidationErrorsSelector(state).size),
-    screeningTitle: getScreeningTitleSelector(state),
-    submitReferralErrors: getScreeningSubmissionErrorsSelector(state).toJS(),
-    staffId: selectStaffId(state),
-  }
-}
-
-function mapDispatchToProps(dispatch, _ownProps) {
-  const actions = {
-    setPageMode,
-    fetchHistoryOfInvolvements,
-    clearHistoryOfInvolvement,
-    clearRelationships,
-    ...personCardActions,
-    ...screeningActions,
-  }
-  return {
-    actions: bindActionCreators(actions, dispatch),
-  }
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(ScreeningPage)
